compat: move cryolab item prototype next to entity
There is a circular dependency right now:

Muluna depends on Titanium
Titanium depends on Krastorio2 Spaced Out
Krastorio2 Spaced out depends on Muluna

I believe the K2SO dependency on Muluna will be able to be removed if
this change is made. K2SO uses an flib function that depends on the
prototype for the `place_result` of an item existing when it is called
on the item.

So, in Muluna, if the `cryolab` item creation is moved to the same place
as entity creation, this should fix the issue and allow K2SO to remove
the dep on Muluna and fix the dependency cycle.

An alternative fix would be to move the `cryolab` entity prototype
creation to the data phase instead of data-updates, but I was not sure
which would be best for Muluna.